feat: DB送信をreserve4でやるよう修正
